app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yookassa==3.2.0
python-telegram-bot==21.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.35